Remove deprecations and move class under `Search\` NS
