+++ UPGRADE-3.x.md
@@ -4,11 +4,6 @@ UPGRADE 3.x
 UPGRADE FROM 3.x to 3.x
 =======================
 
-### `Sonata\AdminBundle\Datagrid\DatagridInterface`
-
-Passing a filter that does not implement `Sonata\AdminBundle\Filter\ChainableFilterInterface`
-as argument 1 for `DatagridInterface::addFilter()` is deprecated.
-
 ### `Sonata\AdminBundle\Search\SearchHandler`
 
 "or_group" option is not set anymore to filters implementing `Sonata\AdminBundle\Filter\ChainableFilterInterface`.
