+++ src/Search/SearchHandler.php
@@ -16,7 +16,6 @@ namespace Sonata\AdminBundle\Search;
 use Sonata\AdminBundle\Admin\AdminInterface;
 use Sonata\AdminBundle\Admin\Pool;
 use Sonata\AdminBundle\Datagrid\PagerInterface;
-use Sonata\AdminBundle\Filter\ChainableFilterInterface;
 use Sonata\AdminBundle\Filter\FilterInterface;
 
 /**
@@ -85,7 +84,7 @@ class SearchHandler
         $datagridValues = $datagrid->getValues();
 
         $found = false;
-        $lastFilter = null;
+        $previousFilter = null;
         foreach ($datagrid->getFilters() as $filter) {
             /** @var FilterInterface $filter */
             $formName = $filter->getFormName();
@@ -103,17 +102,20 @@ class SearchHandler
                     ), \E_USER_DEPRECATED);
                 }
 
-            // NEXT_MAJOR: Remove the `$filter instanceof ChainableFilterInterface` check.
-            if (null !== $lastFilter && $filter instanceof ChainableFilterInterface) {
-                $filter->setPreviousFilter($lastFilter);
-            }
-
-            if ($filter->getOption('global_search', false)) {
+                // NEXT_MAJOR: Remove this line.
                 $filter->setOption('case_sensitive', $this->caseSensitive);
-                // NEXT_MAJOR: Remove the following block.
-                if (!$filter instanceof ChainableFilterInterface) {
+
+                if ($filter instanceof ChainableFilterInterface) {
+                    if (null !== $previousFilter) {
+                        $filter->setPreviousFilter($previousFilter);
+                    }
+
+                    $previousFilter = $filter;
+                } else {
+                    // NEXT_MAJOR: Remove this `else` block.
                     $filter->setOption('or_group', $admin->getCode());
                 }
+
                 $filter->setCondition(FilterInterface::CONDITION_OR);
                 $datagrid->setValue($formName, null, $term);
                 $found = true;
@@ -121,8 +123,6 @@ class SearchHandler
                 // Remove any previously set filter that is not configured for the global search.
                 $datagrid->removeFilter($formName);
             }
-
-            $lastFilter = $filter;
         }
 
         if (!$found) {
